Add new provider setup ui
