+++ src/frontend/apps/dashboard/src/slices/product/scenes/providerDeployments/setupSessionEmbed/types.ts
@@ -1,0 +1,40 @@
+import type {
+  DashboardInstanceProviderDeploymentsSetupSessionsCreateOutput,
+  DashboardInstanceProviderDeploymentsSetupSessionsGetOutput,
+  DashboardInstanceProvidersAuthMethodsListOutput
+} from '@metorial/state';
+
+export type AuthMethod = DashboardInstanceProvidersAuthMethodsListOutput['items'][number];
+export type CredentialsMode = 'existing' | 'new';
+export type SetupSessionState =
+  | DashboardInstanceProviderDeploymentsSetupSessionsCreateOutput
+  | DashboardInstanceProviderDeploymentsSetupSessionsGetOutput;
+
+export type ProviderSetupSessionEmbedProps = {
+  instanceId: string;
+  providerId: string;
+  deploymentId?: string;
+  onComplete: (setupSession: DashboardInstanceProviderDeploymentsSetupSessionsGetOutput | null) => void;
+  onCancel?: () => void;
+  cancelLabel?: string;
+  onWindowOpenCancel?: () => void;
+  windowOpenCancelLabel?: string;
+  onWindowOpenStateChange?: (isOpen: boolean) => void;
+  initialMethodId?: string;
+  hideMethodStep?: boolean;
+  onBackToMethodSelection?: () => void;
+  showMethodStepInStepper?: boolean;
+  hideCredentialsIntro?: boolean;
+  flattenOAuthCredentialsFlow?: boolean;
+  showExternalPreviewSidebar?: boolean;
+  collectAuthConfigDetails?: boolean;
+  initialAuthConfigDetails?: {
+    name?: string;
+    description?: string;
+  };
+  autoStartManagedCredentialSetup?: boolean;
+  onAuthConfigDetailsChange?: (details: { name: string; description: string }) => void;
+  onPreviewCredentialTypeChange?: (type: 'managed' | 'manual') => void;
+  onPreviewModeChange?: (mode: 'managed' | 'manual_existing' | 'manual_new') => void;
+  onActiveStepChange?: (step: 'method' | 'credentials' | 'connect') => void;
+};
